Fixed tons of warnings when using Windows

+++ src/backend/runtime/ImageIO.cpp
@@ -27,10 +27,10 @@ bool ImageIO::save(const std::filesystem::path& path, size_t width, size_t heigh
     EXRImage image;
     InitEXRImage(&image);
 
-    image.num_channels = layer_ptrs.size();
+    image.num_channels = (int)layer_ptrs.size();
     image.images       = (unsigned char**)layer_ptrs.data();
-    image.width        = width;
-    image.height       = height;
+    image.width        = (int)width;
+    image.height       = (int)height;
 
     header.num_channels = image.num_channels;
     header.channels     = new EXRChannelInfo[header.num_channels];
